disabled by default page caching behaviour

+++ backend/src/main/java/com/github/nkonev/blog/listener/hibernate/BlogDeleteListener.java
@@ -9,7 +9,10 @@ import org.hibernate.persister.entity.EntityPersister;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
+
+import static com.github.nkonev.blog.Constants.CUSTOM_PRERENDER_ENABLE;
 
 @Component
 public class BlogDeleteListener implements PostDeleteEventListener {
@@ -23,13 +26,18 @@ public class BlogDeleteListener implements PostDeleteEventListener {
     @Autowired
     private SeoCacheService seoCacheService;
 
+    @Autowired
+    private Environment environment;
+
     @Override
     public void onPostDelete(PostDeleteEvent event) {
         if (event.getEntity() instanceof Post) {
             Post post = (Post) event.getEntity();
             webSocketService.sendDeletePostEvent(post.getId());
-            seoCacheService.removeCachesForPost(post.getId());
-            seoCacheService.rewriteCachedIndex();
+            if (environment.getProperty(CUSTOM_PRERENDER_ENABLE, boolean.class, false)) {
+                seoCacheService.removeCachesForPost(post.getId());
+                seoCacheService.rewriteCachedIndex();
+            }
             LOGGER.debug("sql delete: {}", post);
         }
